+++ DZObjectBuilder/io/__init__.py
@@ -15,6 +15,8 @@ if "binary_handler" in locals():
         reload(data_p3d_odol)
     if "data_xob" in locals():
         reload(data_xob)
+    if "data_anm" in locals():
+        reload(data_anm)
     if "data_tbcsv" in locals():
         reload(data_tbcsv)
     if "data_paa" in locals():
@@ -37,6 +39,8 @@ if "binary_handler" in locals():
         reload(import_p3d)
     if "import_xob" in locals():
         reload(import_xob)
+    if "import_anm" in locals():
+        reload(import_anm)
     if "import_tbcsv" in locals():
         reload(import_tbcsv)
     if "import_paa" in locals():
@@ -52,6 +56,7 @@ from . import data_asc
 from . import data_p3d
 from . import data_p3d_odol
 from . import data_xob
+from . import data_anm
 from . import data_tbcsv
 from . import data_paa
 from . import export_asc
@@ -63,6 +68,7 @@ from . import import_asc
 from . import import_mcfg
 from . import import_p3d
 from . import import_xob
+from . import import_anm
 from . import import_tbcsv
 from . import import_paa
 from . import odol_to_mlod